Put validate_name in shared module and use it in sim engine
For uniformity.
