+++ src/engine/strat_engine/engine.rs
@@ -12,14 +12,14 @@ use crate::{
     engine::{
         engine::{Eventable, KeyActions},
         event::get_engine_listener_list,
-        shared::create_pool_idempotent_or_err,
+        shared::{create_pool_idempotent_or_err, validate_name},
         strat_engine::{
             cmd::verify_binaries,
             devlinks,
             dm::{get_dm, get_dm_init},
             keys::StratKeyActions,
             liminal::{find_all, LiminalDevices},
-            names::{validate_name, KeyDescription},
+            names::KeyDescription,
             pool::StratPool,
         },
         structures::Table,
